Split profit simulator into independent component
