+++ app/src/Sidebar/ScalarSidebar/Buy/profitSimulator.scss
@@ -1,0 +1,148 @@
+.slider {
+  .input {
+    margin: 0.8rem 0;
+    width: 100%;
+    position: relative;
+
+    &::before {
+      position: absolute;
+      content: '';
+      height: 19px;
+      top: 0px;
+      width: 2px;
+      left: 50%;
+      background: black;
+      z-index: 1;
+    }
+  }
+
+  input[type=range] {
+    width: 100%;
+    height: 19px;
+    
+    -webkit-appearance: none;
+    appearance: none;
+    background: transparent;
+    position: relative;
+    outline: none;
+    border-left: 2px solid black;
+    border-right: 2px solid black;
+
+    @mixin track() {
+      width: 100%;
+      cursor: pointer;
+      height: 4px;
+      margin-top: -2px;
+      background-color: #c2e7ec;
+
+    }
+
+    @mixin thumb() {
+      width: 16px;
+      position: relative;
+      height: 16px;
+      -webkit-appearance: none;
+      appearance: none;
+      background-color: #fff;
+      border: 6px solid #009cb4;
+      box-shadow: 0 0 0 0px rgba(82, 104, 119, 0.3); 
+      border-radius: 100%;
+
+      margin-top: -5px;
+
+      transition: all .4s ease;
+      
+      &:hover {
+        border-width: 8.0px;
+        border-color: lighten(#009cb4, 10%);
+        box-shadow: 0 0 0 3px rgba(82, 104, 119, 0.3); 
+      }
+    }
+  
+    /* Track */
+    &::-moz-range-track {
+      @include track();
+    }
+    &::-ms-track {
+      @include track();
+    }
+    &::-webkit-slider-runnable-track {
+      @include track();
+    }
+
+    /* Thumb */
+    &::-webkit-slider-thumb {
+      @include thumb();
+      z-index: 999;
+    }
+
+    &::-moz-range-thumb {
+      @include thumb();
+    }
+
+    &::-ms-thumb {
+      @include thumb();
+    }
+  }
+
+  .labels {
+    display: flex;
+    width: 100%;
+    justify-content: space-between;
+  }
+}
+
+.invest-summary {
+  padding: 1rem 2rem;
+}
+
+.summary {
+  &.profit-loss-sim {
+    .row {
+      .spacer {
+        border: none;
+      }
+    }
+
+    .value {
+      font-weight: 700;
+      color: #00193e
+    }
+  }
+
+  &:not(.profit-loss-sim) {
+    .label {
+      font-weight: 700;
+      color: #00193e
+    }
+  }
+
+  
+
+  .row {
+    display: flex;
+    flex-wrap: 1;
+    margin: 0.4rem 0;
+    
+    font-size: 1.3rem;
+
+    .spacer {
+      flex: 1;
+      width: 100%;
+      border-bottom: 1px dashed #a9b4bb;
+      margin: 0 0.4rem;
+      position: relative;
+      bottom: 2px;
+      
+    }
+  }
+}
+
+.percentage {
+  &-pos {
+    color: #02ae60;
+  }
+  &-neg {
+    color: #db3a3d;
+  }
+}
